Refactor complete project and implement MinimaxAgent
