+++ bot_v_bot.py
@@ -1,16 +1,17 @@
-from dlgo.agent.naive import RandomBot
-from dlgo.goboard_fast import GameState
-from dlgo.gotypes import Player
-from dlgo.utils import print_board, print_move
-from dlgo.scoring import GameResult
 import time
+
+from dlgo.agents.random import RandomAgent
+from dlgo.boards.fast import FastGameState
+from dlgo.gotypes import Player
+from dlgo.scoring import Score
+from dlgo.utils import print_board, print_move
 
 
 def main():
-    game = GameState.new_game(9)
+    game = FastGameState.new_game(9)
     bots = {
-        Player.black: RandomBot(),
-        Player.white: RandomBot(),
+        Player.black: RandomAgent(),
+        Player.white: RandomAgent(),
     }
 
     while not game.is_over():
@@ -22,7 +23,7 @@ def main():
         print_move(game.next_player, bot_move)
         game = game.apply_move(bot_move)
 
-    print(GameResult.compute(game, 7.5))
+    print(Score.compute(game, 7.5))
 
 
 if __name__ == '__main__':
