+++ dlgo/agents/base.py
@@ -1,6 +1,3 @@
 class Agent:
-    def __init__(self):
-        pass
-
     def select_move(self, game):
         raise NotImplementedError()
