+++ dlgo/agents/random.py
@@ -1,3 +1,9 @@
+import random
+
+from dlgo.agents.base import Agent
+from dlgo.gotypes import Move
+
+
 def is_point_an_eye(board, point, player):
     if board[point] is not None:
         return False
@@ -20,3 +26,10 @@ def is_point_an_eye(board, point, player):
             off_board_corners += 1
 
     return friendly_corners >= 3 if off_board_corners == 0 else off_board_corners + friendly_corners == 4
+
+
+class RandomAgent(Agent):
+    def select_move(self, game):
+        candidates = [candidate for candidate in game.possible_moves()
+                      if candidate.is_play and not is_point_an_eye(game.board, candidate.point, game.next_player)]
+        return random.choice(candidates) if candidates else Move.pass_turn()
